Task #3 : First set of changes; shows markers on map.
Loads the earthquake data, creates simplemarks with features. Shows
the map properly.

Offline is true because Google is only allowing about three hits a day
to their maps. I may move it to the MS map provider since it does not
seem to have the as strict a set of limits.