Task #4 : Implement the printQuakes() method
Implement the printQuakes() method in EarthquakeCityMap. This method should use System.out.println() to list each country for which there was 1 or more earthquakes and the number of earthquakes detected in that country. Then it should print out the number of quakes that were detected in the ocean.

Completed... output is sorted as well.  :)