Task #4 : Requirements 8 and 9.
8. Complete the method definitions for the drawEarthquake() method in
LandQuakeMarker and OceanQuakeMarker so that the size of the marker
varies by its magnitude (larger magnitude=larger marker). In addition,
make the LandQuakeMarkers one shape (we suggest circles) and the
OceanQuakeMarkers a different shape (we suggest squares).
Be sure to center them at the x, y position passed in to drawEarthquake.

9. Complete the method definition for colorDetermine() method in
EarthquakeMarker and notice where it and the methods from the derived
classes are called in the EarthquakeMarker?s draw method.

Tested and ready to go.