Task #5 : 7. Implement the mouseClicked() method in EarthquakeCityMap.
This method tests whether the lastClick variable refers to a null object
or not; if lastClick is not null, this click ?de-selects? whichever marker
was clicked on last and so all markers should be displayed. Otherwise,
the method must determine which marker is being selected (if any) and
hide / display other markers so that:

When an earthquake?s marker is selected, all cities within the threat
circle of this earthquake are displayed on the map and all other
cities and earthquakes are hidden. You are given an implementation of t
he threatCircle() method in the EarthquakeMarker class.

When a city?s marker is selected, all earthquakes which contain that city
in their threat circle are displayed on the map and all other cities and
earthquakes are hidden. You are given an implementation of the
threatCircle() method in the EarthquakeMarker class.

Unit tested and ready to go.