Task #6 : Implemented keypress handling.
Added code the keyPressed() method to caputure keystrokes and
convert them to floats for the number keys to for use as a filter
on earthquake display. Still need to implement the actual hide and
unhide for earthquakes based on the filter. Unit tested by displaying
the captured values to the console.