Task #6 : enhanced the keypress handler and added commments.
This version of keypressed reacts to the number keys as discribed before
unless the shift key has been pressed. The shift key toggles modes; if
"shifted" the we show quakes from with magnitude from n to n+0.999; if
unshifted (the starting state) we show quakes with magnitude >= n.
toggling the shift key will flip between the two modes while maintaining
the existing n value. Any other key resets n to 0.0 and shows all quakes.

The output to the console has been improved as well. It now states if
it's showing all quakes, all quakes greater than n, or quakes from n
to in +0.999.

Unit tested and ready to go.